Move observer map creation to orchestrator
